Update vendor support to reflect v2 non-task
Vexxhost and RunAbove both support v2 of glance, but do not require
the task interface. Update the erroneous information. Also, there are
two additional regions for Rackspace that we did not list. We don't
list the Rackspace LON region because it does not work like the others
and we do not know how to support it.

Change-Id: Ib155d00d1d6bf7b2e5bcf4b868c268561e198611