Split security group list operations
Security groups can come from either nova or neutron or not be
available by the vendor. This uses a new flag from os-client-config
to determine where we should get the list from. Also, a new exception
is added for features that are not supported.

This required a minimum version bump in os-client-config.

Change-Id: I4474341ebc255700a06247846c6350981437e8ae