Fix delete_server when wait=True
Functional tests were running with wait=False, which hid that we
are not handling the expected NotFound error after delete. So we
refactor delete_server to use get_server and be much simpler. Note that
we call the ServerGet task directly since we already have the ID that we
want to make sure we deleted, no need for listing at that point.

Adding unit test coverage to help avoid these problems in the future as
well.

Change-Id: Ib78aef75a5ad3f75bc8ee2b81078d30cc2fc76db