Change overriding defaults to kwarg
Forcibly changing the global defaults isn't the nicest thing to do
and turns out to be a pain during tests.  Make overriding defaults into
an arg to OpenStackConfig.__init__().

OSC is the only known user of set_default(), once this is released and
in global-requirements OSC an be updated then set_default() can be removed.

Change-Id: Iddbc66398e89f06f1d665d7b0ef243bc786c8e36