Common head method for proxies
This change adds a common HEAD method for proxy methods to use. It
currently supports calling HEAD directly on a resource (which is done in
at least one place, to get account metadata in object_store), as well as
the likely more common case of requesting HEAD for a particular resource
via the `value` parameter (such as getting container or object
metadata).

Change-Id: Icec939d40c6028c245c9503cbdfecfa801c265ef