Add Neutron/Nova Floating IP delete (i.e. deallocate from project)
Some clouds out there are still running nova-network but we want to
provide support for neutron specific features and concepts like ports,
networks, subnets, etc.

This change is part of a set that adds neutron support to existing
floating IP-related functions, hiding that behind resource-oriented
methods.
For instance, at high level, end-users can now request that a public IP
is assigned to an instance without worrying about the specific service,
procedures and protocols used to provide that feature in the target
cloud.

Change-Id: I699d65ef3f9d917e14b4046d5ee4e6763e2ecc74