Accept intermediate path arguments at proxy
We need a way to get what amounts to "path_args" in the resource layer
through via the proxy layer. This change adds support in the BaseProxy
class for sending the path_args through, and applies it to several
object_store calls that need the container name to come through while
working with object operations.

Change-Id: I9955330078bca9a6b1992b74b0c14822d3e44957