Decorator for functional tests to check services
Spawned from https://review.openstack.org/181607, we may need the
ability to know whether or not certain functional tests can be run. This
change looks at the service catalog via a ServiceFilter to see that an
endpoint can be found for the particular service being tested. If one
exists, the test will run. If one doesn't the test will be skipped.

Change-Id: I7dca09804fe8663a316b807dbf9f5a77c77538c2