Always refresh glanceclient for tokens validity
Glanceclient doesn't refresh tokens on its own. Work around this by
always making a new glanceclient with a potentialy new token if the
existing token is near expiration. This adds some overhead to our use of
glance but that is preferable to having glanceclient break because its
token has expired.

Change-Id: If57531a72eb90ee7bc6e67905ddfd5bda9bb6f1b