Rename messaging module to message
Looking at the project names [1] it's clear that the official names don't end
in 'ing' (the only exception being Networking). The SDK follows the official
names as our module naming convention.

Hence we should not append 'ing' to module names whether they have an official
name in that list or not. We need to rename the messaging module to message.

[1] https://wiki.openstack.org/wiki/Documentation/Conventions

Change-Id: I2a433cb72341e2660e33cd3f946e2f22eed19622