Add list of image params needed to disable agents
Some clouds have an in-instance agent to handle things like online
password resets and other such activities. When building and uploading
images, it's often advantageous to not install such an agent and instead
handle such things via config management... but doing so requires data
to be set on the image itself.

Change-Id: I5b7c9d72fd2d49890bc466d7dd22a3cb9595f670