Provide a helper method to get requests ssl values
There is a weird logic around the interaction between how the openstack
things all accept cacert and verify/insecure parameters and how requests
wants them. Rather than spreading the parameter combining logic across
the universe, put it here.

Note that this inverts the usual requests logic in that !verify will
override the presence of a cacert value and cause verification to NOT
occur.  This is intended to become the normal mode of operation for
OpenStack clients.

Change-Id: I3c76d9a10e6e5d60a593ceefc87dafdc6857d9c6