Add support to get a SwiftService object
This adds a property of the cloud, swift_service, for the purpose of
enabling uploads via SwiftService's more advanced segmenting
capabilities.

In writing tests, it became clear that cloud.swift_client was leaking
exceptions, which we don't want it to do, which is why we add exception
handling to that method.

Change-Id: I045266e6d5f0b6bed3100905ada6ab35fbb06987