Cast nova server object to dict after refetch
When we assign a floating ip to a server, we re-fetch it from the cloud
so that the metadata is all correct. However, we weren't doing
obj_to_dict on it like all of the other code paths that return out of
create_server - which means that it bombs out on
get_hostvars_from_server if subsequently passed to it.

Change-Id: I3f69f525b7d57ec91f70b77cdbd62aed9816f17d