Claim messages and delete messages
Implmenting claim messages and delete messages necessitated a refactoring
of the proxy interfaces. They were ugly and cumbersome. The information
was pushed into the objects. This comes at the expense of a bit of
validation when creating messages but it's minor.

The new design is more in tune with the hypermedia nature of the Message
Service API. Before I was fighting against it, trying to parse out ids.
That's a recipe for brittle code. Much better to make use of the hrefs as
they're intended to be used.

Change-Id: Ib14632a2a05539eae060771f46e0a313a652903f