Compare message data outside of assert_called_with
The sort order of message data in a dictionary is sometimes inconsistent
between what is passed in the system under test and what is created in
the test method for comparison. Instead of having assert_called_with do
the comparison, test that it is called with anything at all, and then
breakdown what was sent in the call and check the dictionary contents
independently. This will fix the spurious test failures under Python
3.4.

Change-Id: I0182c9d8aaecf28db24b59ef3039a9c873ee2b5c