Refactor verify_create in proxy tests
This is the first of several changes to cleanup the proxy tests and
their duplicated verification mocks.

Note: this is slightly changed for the test_messages_create and
test_messages_claim tests in TestMessageProxy. They aren't using the
BaseProxy methods so they need to be implemented slightly differently.
It's not entirely clear how they were actually working before because it
was mocking out a method the call doesn't go through.

Change-Id: I16e28b3126a474fd14a9fdde433620009a83bf29