Don't emit volume tracebacks in inventory debug
It's ok if a cloud doesn't have volumes, and it's a question we can ask
of the cloud. Now, keystone tells us about it by way of throwing an
EndpointNotFound exception. But it's a normal logic exception, not a
"this is unexpected" exception. So we can handle it, log a sane error,
and move on, rather than logging the full traceback.

Change-Id: If6fc1ab0bb8247e0fdbd1920aa0b076ac1707cba