Use json for in-tree cloud data
In preparation for sharing the default and vendor data with other
projects, potentially even non-python ones, move the data into json
format, which is slighly less exciting to read, but has more widespread
standard library support. The user-facing config file will still be in
yaml format, because that's easier on the eyes and it's expected to be
read and edited by humans.

Continue to accept yaml everywhere, because an end user may have dropped
a yaml config file into a dir somewhere, and that's fine.

Change-Id: I269d31e61da433ac20abb39acdde0f9f9fe12837