Add get_server_external_ipv4() to meta
It turned out we can't rely on get_server_public_ip to get an
externally reachable IP of servers because not always floating IPs
are used and not always are what the caller is looking for.

get_server_external_ipv4() tries hard to provide an IP address of the
given server that can be reached from networks that are external to
the cloud.
Unfortunately from some clouds (or for very particularly configured
servers with multiple interfaces) it could not be possible to
reliably determine those IP addresses.
Those clouds are corner cases that are not supported by
get_server_external_ipv4()

Co-Authored-By: Monty Taylor <mordred@inaugust.com>
Change-Id: I099193437d0f45cfda78923349e500e9b2e0e053