Rename volume to block_store
In line with other renamings we've made, the volume service is really
the "block store service". It's still called volume in the service
catalog, which is why it was originally named that, but block storage is
how it exists in projects.yaml as well as how it's actuall referred to.

Change-Id: Ic97dfc83561844dd3dcbd569b272f8f1a853f436