Rename project to openstacksdk
Prepare for the next release and set the package name to openstacksdk
and drop the python- prefix. This package name has already been reserved
on PyPI.

This was initially started by a discussion at the summit around
OpenStackClient making the same move to strip the python- prefix
(they haven't completed it, but have reserved the PyPI name).
It was further discussed on IRC and I believe in a meeting that
there is no benefit to having the python- prefix here. There is
no name collision we're avoiding with it, and it provides
nothing extra by being prefixed as such.

Change-Id: I1145883a2f896769143a3fcc6e43468d6e590e11