Return True/False for delete methods
This changes many (not all) of our delete methods to have consistent
return values: True if the resource was deleted, False if it was not
found for deletion.

The methods being changed here either returned nothing at all, or
they raised an exception on resource not found. The ones that raised
an exception no longer raise this exception, and instead return False.

We also make sure that a debug log message is output for a delete
request that could not find the resource.

Change-Id: I0ee4654dc6f68b475179d36002fe9db713e26290