More selective logging
Having the SDK configure the root logger is problematic as it
not only enables logging in the SDK but it also enables
logging in the users app and logging in other libs unrelated
to the SDK as an unintented side-effect. We should only log
for the openstack module.

We also want to be more selective about what stream to log
to so we provide the option to log to stdout or stderr.

Change-Id: I82ef867e182ebe4f29e21380268a3691e3b83494