Claim no messages correctly
The Message Service will respond with a 204 and no content in
the body when there are no messages to claim. The transport
layer doesn't like that and we have to correct for it here.
Ultimately it's a bug in the v1.0 Message Service API.

Users shouldn't have to deal with an InvalidResponse just
because there are no messages to claim. That's a perfectly
normal state for a queue to be in.

It appears to be fixed in v1.1
https://wiki.openstack.org/wiki/Zaqar/specs/api/v1.1#Changes_from_v1.0

Change-Id: I5e8a181ba8f633a387cb8e237e9bf9d90ade680c