Check results of find before returning
Resource.find was returning values that may not have actually matched
what was attempting to be found. When only one item was found, it was
returned regardless of match. This can be seen if you only have one
server, then do compute.find_server with both a real server name and
some fake name - you'll get that one server back every time.

This test removes the "test_nameo" test, which was only working because
of this bug.

Change-Id: Ied12c5a3badc3d6c5f47d3b2d8b86e1dc9d1d1fc
Closes-Bug: 1419576