Modify secgroup rule processing
It turns out that Neutron can accept None for a rule protocol to
represent all protocols. Nova has no similar concept for a single
rule, so let's throw an exception.

Also, Neutron relaxes the rules on accepting None for both port
values to represent the full port range for TCP and UDP. For ICMP,
it is the same as -1.

Change-Id: Icc408586804e2a538399a2aa3f740b97bb4c5246