Add support for configuring region lists with yaml
yaml supports encoding lists in a manner that is cleaner than comma
separated lists. While the old commas-in-region_name will still work,
add a 'regions' option that can be used to configure lists of regions.
Remove the comma-separated-list from the docs so that people don't try
to use it - even though it will work.

Change-Id: Ieb0aedb9c03fd26e644e9ba733b935f2c69daaf0