Allow use of admin tokens in keystone
For bootstrapping purposes, an admin uses an "admin token" which is a
statically configured secret in the keystone config file. If one has
used this, there are things one cannot do, such as query the catalog.
That's ok - because what you're doing is bootstrapping the contents of
the catalog.

In order to enable that, we need to work around the normal "just infer
the right thing from the catalog" logic, because there is no catalog.

Also - it turns out that direct use of the v2 and v3 clients allows one
to just pass in a Session, rather than needing to query the session for
the auth_url to pass in. This does not hurt the non-admin cases.

Change-Id: I7e15214a589dd3cb270f4835e074dc13cab3e529