Add or change timestamp suffixes to "_at"
The great majority of our time related properties on resources are named
"something_at". That started because the body values in responses in a
number of resources are actually named that, and we seem to have
converged on that naming scheme over "something", as that could in some
cases appear as a boolean (updated? Is it true, is it the time, etc).
This change corrects a few that were un-suffixed or that had a different
suffix. We should now be consistent on time related properties.

Change-Id: I4bed7344a156df5dfce62af5c28bbffad3b5857e