Fix identity domain methods
Some bugs made it through the review process. The delete_identity_domain()
method was calling an invalid update method. It also should only accept a
domain ID, not name, since converting from a domain name to ID would require
listing the domains (an expensive operation we are trying to avoid).

Change-Id: Iccc89f133b4122d88bc77f72d89ae6c12eff1eb9