Fix set_default() when used before config init
Using set_default() before initializing OpenStackConfig would cause
an error since the _defaults dict would still be None and not an
actual dict. This corrects that by calling get_defaults() to make
sure it is initialized properly, and also adds a warning to note
that the method is now deprecated.

Change-Id: I81803c680b614f9bee47c6f69a4efffa638dcebc