Always use a fixed address when attaching a floating IP to a server
When using Neutron, if the caller didn't specify a fixed_address when
assigning a floating IP to a server, automatically pick the fist IPv4 address
assigned to the fist port of the server.

This should fix the DreamHost use case: DH uses Neutron and automatically
assigns an IPv4 and an IPv6 fixed address. Even if it doesn't make sense to
attach a floating IPv4 address to a fixed IPv6 address, OpenStack returns an
error if we don't specify a fixed address when attacking a floating IP to a
server port when multiple IPs assigned to that port.

Change-Id: I4b91bf29366c4d8b5277d0d97cd9571770252961