Change functional testing to use clouds.yaml
Using clouds.yaml, instead of sourcing openrc, we can have better
functional testing by distinquishing between an admin account
and a normal user account.

We cannot both source openrc AND use clouds.yaml as this confuses
the inventory functional tests, and is not a recommended way of
utilizing o-c-c, anyway.

In order for these test to work, a bug fix was needed in
os-client-config, so the requirements was bumped to 1.6.2
or higher.

Depends-On: I6458969b45f5534e1172b9f8ba598d7c536a59b5
Change-Id: If459afba3cf61b3c7f641c79494cd53a682666a4