Revert "Revert "Use the correct auth_plugin for token authentication""
Name the auth type needed for using an admin token "admin_token". This
hack can be removed with keystoneclient and/or keystoneauth make a
release with the admin_token entrypoint defined.

Naming it admin_token further reduces the conflict with OSC, which
should make the patch to work around OSC having different arguments
unnecessary.

This reverts commit 0e3117cb01d201f403be72cbb1218b9eec69a541.

Change-Id: I079c7c61b2637ded73542876cb1378a0731f8631