Only log text strings in requests
As was exposed by object_store.create_object, we were attempting to mix
text and bytes in log messages, which was causing _logger.debug to choke
when combining the parts of the message, specifically the --data piece.

This is an issue for two reasons:
1. It's binary data in general
2. It's a very large amount of binary data

The fix is to not log byte strings that come in as `data` unless we can
decode them to ascii. This will allow us to continue logging some data
such as auth credentials, but we won't be logging megabytes or gigabytes
worth of binary data - such as objects being uploaded to Swift - into
the log file.

Change-Id: Iba49552b8ebc58a46cedf826751193a1fa7feb11
Partial-Bug: 1485054