Properly pass on Container in path_args
Once we moved object_store from its own form of proxy to using the
BaseProxy methods, we lost proper handling of the container during
object-related operations. We need to pass on container.name, not the
full container object, so when instances were passed in a URL would be
created with the repr of an actual Container, which clearly doesn't
work.

There's an additional safety check in create_object to make sure a
container is passed in before trying an operation that won't work and
then raise a confusing exception.

Partial-Bug: 1485054

Change-Id: I85bb6f3b09a9617395dba0fd9becb62a8c4d7e68