Fix find_extension for Network and Compute proxies
The Extension resource for both Network and Compute were missing two
pieces that would help their find_extension methods work: for one, the
`alias` is actually the ID of an extension, and no ID value was set. If
we knew what the ID was, we could find it (or not find it) directly.
Before that can even be done, we would need to be able to get it by ID,
and we didn't have allow_retrieve set on either of those resources.
Allowing Extension resources to be retrieved by ID causes them to now be
found.

Closes-Bug: 1456317
Change-Id: Id2a2071e9649edca78b2a5455a27e45443d6e8c9