Get container off of an Object if its passed
The previous object operations on object_store proxy were too strict on
requiring a container argument to be passed, and in some cases they were
too loose.

When an Object is passed, look to see if it has a container attribute
already set and use it if so. If that doesn't work out, look at the
container argument, which that that point would have to be specified. If
we get a container out of that, use it. After this, we won't be able to
do any object operations with a container, so raise ValueError
specifying that a container has to be provided.

Change-Id: I6b587831321998245815fccf8f822025e83141c5
Closes-Bug: 1487147