Return keystoneauth plugins based on auth args
We know all of the things that we need to know to return an appropriate
auth plugin from keystoneauth based on the auth parameters. This also
introduces a hard-depend on keystoneauth, which should be fine since
keystoneauth itself has a very low dependency count.

We'll also use ksa to help validate auth parameters when we're doing
the config processing.

Change-Id: Ia1a1a4adb4dcefed5d7607082e026ca7361f394d