Allow configuring domain id once
In the most common case, a use has one and only one domain id that they
care about, and it's associated with their user and project. Instead of
making them set it for both user_domain_id and project_domain_id, allow
for domain_{id, name} and then fill in any missing values for
{user,project}_domain_{id,name} with the given value.

This is mainly because I wind up with config files looking like this:

        user_domain_id: d0919bd5e8d74e49adf0e145807ffc38
        project_domain_id: d0919bd5e8d74e49adf0e145807ffc38

Which offends my tender sensibilities.

Change-Id: I12342dfa9f1b539a3fea5dd8874c42d027c59739