Properly handle os- prefixed args in fix_args
Prior to this any arg that started with 'os' had the 'os' prefix chomped
including the next character. This meant that any arg like 'osmosis'
would've been replaced with 'osis' which is clearly wrong. Instead we
want to treat the OpenStack prefix of 'os-' or 'os_' as the special
thing so check the next character is correct before chomping.

Change-Id: Id12a92adf63d896f7aa5c0e391abd299c4ce3331