Store the inner exception when creating an OSCException
Masking exceptions in Python is an anti-pattern.  Etc.

However, since shade uses a base class of OpenStackCloudException
for all of its exceptions (and, for the worst offenders of
generalized exception masking, simply uses that exception itself),
we can store the original exception for clients to access if needed.

This does allow client exceptions to leak out of shade to its users,
but in a very controlled manner where we can make it clear that
this should not be used for flow control, but for informational
purposes.

Change-Id: I33269743a8f62b863569130aba3cc9b5a8539aa0