Remove many redundant debug logs
Now that shade stores inner exceptions in the OpenStackCloudException
constructor and logs them, there is no need for logging the same
exceptions at debug level in the library.

This patch removes all such log entries that appear to add no additional
information beyond what is in the exception.  Many messages remain
which add information not in the exception -- their accompanying
exceptions should be updated and then the extraneous log lines
should be removed.

Change-Id: Idcdbfa77c68260d171e78f6ada4ee5a46518061c