Fix baremetal port deletion
Shade's node deletion logic removes ports one at a time in
the requested order, however that logic has been silently failing
to the API for some time now as we're inadvertently returning the
entire port object back to the API as the port we wish to delete
instead of the port UUID.

Changed to return the port UUID value to the delete call.
Added code to unit test to ensure that correct action is taken.

Closes-Bug: #1500063
Change-Id: I01c16990dc66988fab29c41ec4b4e6094213b2ca