Provide short-circuit for finding server networks
Doing a search for the networks on every server connect can be
expensive due to the port list - even with caching. Instead, allow
for a configurably overridable networks to be found once and cached
aggressively, then use that network name to match in the nova networks
dict. If that can be found, awesome, we all win.

Change-Id: Ic3345e7fb7e428227a9465bd07077e67d4f93244