Put in override for Rackspace broken neutron
Rackspace puts neutron in the service catalog but it does not work (404s
the endpoint) While it's easy enough to try/except around network calls,
it's also a known fact and can be communicated clearly. Allow things
that want to short-circuit read the value from the config and react
accordingly.

Note: This is not a value going in to the defaults, because the fact
that this is happening at all is highly strange and hopefully something
that can be solved by collaborating on future iterations of DefCore.
However, communicating clearly that it's a known issue is helpful for a
class of users.

Change-Id: I506f4dd397608f2feb6fbda48d297d283312a169