Fix create_image_snapshot
This method from novaclient returns an image id, not an object
as was apparently assumed.  Before returning, perform an image
lookup so that we can return a Bunch() object of the image.

Note that if the provider does not immediately make an image
available for lookups, this will fail and the user will have
no record of the image id.  Notably, because it worked with
the image_id until the image was built, nodepool was not suceptible
to this problem.  I am unsure whether there are any clouds that
might behave in this way.  However, this code matches what we do
in shade with servers.

Change-Id: Icaff5d8e66a2458817fcbced16014efbc5ff33b8