Optimize network finding
clarkb was right about the port listing code being unneccesary now
that we have get_external_network. All of the logic that happens there
is also more efficient and easier to override via config. Skip doing
a second pass at ports.

Also, move floating ip finding earlier - if the server has one, that's
the IP it wants - it's the only reason to use one of those.

Change-Id: Ie09c0046aded3646a42cc9714f1fb76375e5e099