Invalidate image cache everytime we make a change
The image cache is already set up to only cache a steady-state list. If
we simply poke the cache after every time we make an upload or image
change, the existing cache code will cause the cache to stay unfilled
until the new change hits steady state again. This should make turning
on cache do the correct thing.

We want to follow this up with a get bypass of the list, but this should
keep our API counts to a reasonable level.

Change-Id: I47cdfb5af1ba5c56c4005289b9d6b1ecde936978