Split get_hostvars_from_server into two
There are really two different things going on here. One is figuring
out essential information "what IP do I use to talk to this server"
The other is fleshing out some additional information that is useful
for ansible inventories. The ansible inventory info is expensive when
that's not what you're doing and when you're not using shade-level
caching.

A follow-on patch will make the create/get/list server calls use the
new function always.

Change-Id: Ia1de78d26c708ce6fe3205a9484cb16a92360890