Provide option to delete floating IP with server
Although the only real reason you should need a floating IP are for
the cases where you want an IP with a lifespan independent from the
server, there are clouds that are insane and require a floating IP
to be created with every server. When a person is deleting that
server, we should ensure that the floating IP is delete. However, it's
impossible to infer intent about whether the floating IP wants to stick
around, so it has to be a flag.

We will default to False because deleting things that the user did not
expect to be deleted is surprising in a bad way.

Change-Id: Id90bdc819ad8f1b04acc1b35788786d23f4e7666