Add option to floating ip creation to not reuse
Although reusing available floating ips is potentially a more
efficient use of resources in the cloud, it's a more expensive
amount of API calls. Additionally, if the user knows that they
are deleting floating ips with servers, then attempting to
reuse is just going to be a waste of effort.

Co-Authored-By: James E. Blair <jeblair@linux.vnet.ibm.com>
Change-Id: Ibfb7583232b03bcf29df229862f6da97170aa845