Add create/delete for keystone roles
This completes the replacement of: Ifd74cbcef9dd0f3531912995da8ad7b75dd19d44

Add create and delete methods for roles. Note that there is no
update method (the review we are replacing had an update method).
This is because keystone v2 does NOT have support for this, only
v3.

Change-Id: Ib037484c38673f10bd2b2fe2df5ca87056d06e6d