Push filtering down into neutron
All of the neutron list calls support passing filters to the cloud for
server-side filtering. Make use of that so that for things like port
lists in nodepool we don't wind up listing a bazillion ports all the
time.  This should reduce stress on the cloud. The number of API calls
we make will be the same, but payloads smaller.

Change-Id: Ib134c307c135aef5614f8032d965e61ca7242323