Clean up cache interface, add support for services
We just added an unreleased interface method to the CloudConfig object -
but maybe that should have been more aligned with dogpile words.

SO - change the docs to reference the dogpile words and add support for
that, while keeping backwards compat support for people using max_age.
Also, do the -/_ transform on the cache config like elsewhere.

Then, while we're in there, add support for per-service cache timings.
We need this in nodepool and shade is adding support, so ability to
configure it will be important.

Change-Id: I31190a31ab0b79fc080db3611c0cd584076387d4