Make designate record methods private for now
The API for designate is not something we can commit to right now. It
turns out that v1 has Domains and Records and v2 has Zones, RecordSets
and Records. Figuring out an API design that can handle both is not
going to happen before 1.0, so mark the current Record API as private
so that we can add something we can commit to later.

Change-Id: Ia2611f19d04a2d7a51d85ac99f19ed4054972085