Allow for templated variables in auth_url
Some clouds have auth_urls per region, which means putting a static
auth_url in the config file isn't going to work. Allow {} substitutions
in the auth_url so that other elements of the auth dict can be injected.
This will not solve _everything, but it will solve all of the currently
identified issues, and should be upwardly open if we need something
more complex later.

Change-Id: I6107d0669734647cfa0bef118bdf7739949991ad