Make raw-requests calls behave like client calls
NovaGetUrl returns a tuple of response/body, but everything else returns
the actual object. If we're going to make raw calls, we should handle
them in such a way that the rest of the code doesn't konw the
difference. This also means that we have the opportunity to look at
response codes and throw exceptions if they aren't 200.

Change-Id: I0bab5056b5ba234f5509b56a716ce426e6a2c604