Fix keystone domain searching
Domain searches is a bit different in that it pushes down the
filtering to the keystone client. This is fine, except that it
didn't work as the filters were not actually passed.

This change passes the filters, and changes the search_domains()
API to accept a 'filters' parameter to be consistent with our other
search methods. It previously accepted only keyword parameters rather
than just a filter dict.

Change-Id: I61c8591493c5d10de38a6541a9859be2950d60a6