Don't assume pass_version_arg=False for network
Excitingly, one of the places where we hard-code a workaround for
a differently behaving client is not needed. The Client in neutron
that takes a version arg is in neutronclient.neutron.client. Also,
pass the version arg as positional when we pass it, as some of the
clients (neutron) require it, and all accept it.

Change-Id: Ifcbaab782173b95a678af0b2792a1194b198b687