Remove designate support
It was broken, thus it was impossible that anyone was actually using it.
We do not have a plan for v1/v2 compat support yet, and it's a hard one.
SO - rather than leaving a broken thing in the tree which was not
possible to use, or fixing it and then having a half-baked thing which
people might actually start using, pull it out pending someone stepping
up and caring about it.

Change-Id: I390494fee8c6c7fb5ddde2ce47e7eb598ad5e08b