Work around a bug in keystoneclient constructor
keystoneclient bug #1513839 means that even when you proerly pass
a Session (like we do) to the discovery constructor (what you'd be
calling if you were passing "pass_version_arg = True") ksc fails
because you didn't send a URL.

We _HAVE_ an appropriate URL that we can pull from the Session. So
until ksc learns how to pull the URL from the Session itself, do it
for them.

Change-Id: I38eb4cfa750fab5196b86989c2cd498d41bf37ac