Add authorize method to Connection
The ability to know your application is going to fail due to a
credential issue *before* getting into the application and attempting to
make some API call can be useful. This change adds Connection.authorize,
which just does the same thing we do implicitly. It's entirely optional
and is documented as such.

Co-Authored-By: Qiming Teng (tengqim@cn.ibm.com)
Closes-Bug: #1513200
Change-Id: If86f308da909c4ff31f37294a000c6972c947182