Add support for legacy envvar prefixes
In trying to move the legacy clients to do their config via
os-client-config, many of them had prefixes like NOVA_ or GLANCE_
in their shells. It's pretty easy to support optionally passing those
in so that the transition to OCC is less of a change and we can
treat the deprecation cycle of those features as a different topic.

Change-Id: Ic819c0790989fa034db03d83535ee7b70aaacc3a