Fix name of the object-store api key
There is only one value for this, and it's not consumed in the process
of making a swift client, so the chances that anyone set it
to something else are pretty much nil. However, for completeness we
should make it the right name, as "object-store" is the service key
name for swift, not "object".

Change-Id: I395c1c44a2f50996b61dff22e07149b8dd13eda9