Handle cinder v2
It turns out that cinder v2 has a service_type of volumev2 because
nobody thought of the children. But that's ok - we actually care about
user experience around here. SO - take the sane approach and return
service_type = volumev2 if service_type == volume and volume_api_version
== 2. This way user code can all safely say "please give me the endpoint
for the volume service" and can use the api_version parameter to specify
which version they want.

We should all possess righteous indignation about this patch.

Change-Id: I15fc5ddd92345d78b6928f11a8d77cecd0427f7d