Add user group assignment API
Add methods for adding and removing users to/from groups and
checking for membership.

This also updates the task_manager code to look for more types
that we don't want to munchify because one of the new keystone
API methods being called returned a bool value.

This replaces Ida3cff3acdc1406c5e6d61500766a292565191fc

Change-Id: Ib34c116010312ed26b042621fcf2e7b5b774424f