Add find support to BaseProxy
All other actions on the Resource class that are used in Proxy
subclasses had been exposed in BaseProxy except for find. This change
implements a simple `_find` in the base, which brings it in line with
how everything else is done, and is the last piece to get knowledge of
`self.session` out of proxy calls.

Additionally, while moving the proxy implementations to all be the same,
it's a chance to unify the find tests in the proxy. This change
eliminates the need for there to have been two test methods and makes
the tests look like how they're done for every other action.

Change-Id: Ibea01bed4e4ecd76d96021f78db0116012448984
Closes-Bug: 1517149