Fix a 60 second unit test
The no-addresses test was waiting for delete timeout and was therefore
sticking around for 60 seconds for a sideeffect we did not care about.

Add a second get return value so that the get in delete returns without
a wait-poll loop. Also, set the list server interval to 0 so that we
done have a bunch of tests with a single 6 second wait.

Change-Id: Ia62ace386a183892f9446aee4e9d4cbbee528dec