Make delete_server() return True/False
To be consistent with our other delete API methods, make delete_server()
return True if the delete succeeded, or False if the server was not
deleted because it was not found.

This also combines the functional testing of create_server() and
delete_server() so that tests are more efficient by eliminating an
extra instance creation. This is in anticipation of adding more
tests that will create server instances and there are only so
many resources available in the test environment.

Change-Id: If1b1a3886df60fb064682caa23113f91b78a4090