Fix cluster action api invocations
As an attempt to be more comformant to the api-wg guidelines, Senlin
API has changed the interface for action invocations. The request method
is now POST instead of PUT and the URI now contains a plural 'actions'
rather than 'action'. This patch fixes the SDK resource to adapt to this
change.

Change-Id: Icbc12db311da2b73e92a5d1d31d94f7564ecb462
Depends-On: Ifa5d57a8d1a6f96f535f95a376726b1a8c066630