Provide a better comment for the object short-circuit
A previous commit was landed over a -1 from a core reviewer that was
concerned about the terse and flippant commit message that accompanied
the commit. The substance of that objection was that there was no good
explanation for why the change was made. Indeed, that why is likely
useful for people looking at the code too, so add a more complete
comment on the code in question.

Change-Id: I1ff777393502edf4ef3ed595e47bfcf73be0e5e8