Remove incomplete CDN code
OpenStack Poppy, the CDN service, was going to be added in here, but it
never ended up being completed. This change removes an empty proxy class
and a few template documentation mentions. This should be added
eventually, but should not exist right now since we're not going to be
fleshing out the details between now and 1.0 anyway.

Change-Id: I529c4901e11161158ce5c35df3bff2ecbbb3091a