Remove the Metric proxy
This proxy class is not currently useful in any usable way so it just
pollutes the namespace of loaded proxy classes. It was added
to get the class started, but was not completed. Rather than offer a
proxy that only tells you what it can do but can't do anything, it's
removed. The resources that would populate a proxy class are left alone.

Change-Id: I30dbff0dbba26633d3b02ee5f0751b4224cef8f4